# api/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorDto": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message",
          "status"
        ],
        "title": "ApiErrorDto",
        "type": "object"
      },
      "ApiErrorEnvelopeDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelopeDto",
        "type": "object"
      },
      "Body_PublicApiController_submitDocument": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Document to process.",
            "title": "File",
            "type": "string"
          },
          "llm_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional LiteLLM embedding profile.",
            "title": "Llm Profile"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_PublicApiController_submitDocument",
        "type": "object"
      },
      "EmbeddingsData": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          }
        },
        "required": [
          "job_id",
          "result"
        ],
        "title": "EmbeddingsData",
        "type": "object"
      },
      "EmbeddingsResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EmbeddingsData"
          }
        },
        "required": [
          "data"
        ],
        "title": "EmbeddingsResponse",
        "type": "object"
      },
...
      "JobDto": {
        "properties": {
          "chars_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chars Used"
          },
          "chunk_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Count"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "llm_transform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LlmTransform"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "$ref": "#/components/schemas/ResultAvailability"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "tokens_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Used"
          }
        },
        "required": [
          "job_id",
          "state",
          "file_name",
          "created_at",
          "result"
        ],
        "title": "JobDto",
        "type": "object"
      },
      "JobListData": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobDto"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "has_more"
        ],
        "title": "JobListData",
        "type": "object"
      },
      "JobListResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobListData"
          }
        },
        "required": [
          "data"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobDto"
          }
        },
        "required": [
          "data"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobResultData": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "state": {
            "const": "succeeded",
            "default": "succeeded",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "result"
        ],
        "title": "JobResultData",
        "type": "object"
      },
      "JobResultResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobResultData"
          }
        },
        "required": [
          "data"
        ],
        "title": "JobResultResponse",
        "type": "object"
      },
      "JobState": {
        "enum": [
          "queued",
          "processing",
          "succeeded",
          "failed"
        ],
        "title": "JobState",
        "type": "string"
      },
...
      "ProfilesResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProfilesData"
          }
        },
        "required": [
          "data"
        ],
        "title": "ProfilesResponse",
        "type": "object"
      },
...
      "SubmitDocumentData": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "state": {
            "const": "queued",
            "default": "queued",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "SubmitDocumentData",
        "type": "object"
      },
      "SubmitDocumentResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SubmitDocumentData"
          }
        },
        "required": [
          "data"
        ],
        "title": "SubmitDocumentResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Optional local DocRunr API key",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local API for document extraction and optional embedding generation. Use a Bearer token when API_KEY is configured.",
    "title": "DocRunr Public API",
    "version": "0.1.2"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/documents": {
      "post": {
        "operationId": "PublicApiController_submitDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_PublicApiController_submitDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "JOB_NOT_READY"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "PAYLOAD_TOO_LARGE"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Upload document",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/jobs": {
      "get": {
        "operationId": "PublicApiController_listJobs",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List jobs",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/v1/jobs/{jobId}": {
      "get": {
        "operationId": "PublicApiController_getJob",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "title": "Jobid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "NOT_FOUND"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get job",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/v1/jobs/{jobId}/result": {
      "get": {
        "operationId": "PublicApiController_getJobResult",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "title": "Jobid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "markdown",
                "embeddings"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/JobResultResponse"
                    },
                    {
                      "$ref": "#/components/schemas/EmbeddingsResponse"
                    }
                  ],
                  "title": "Response Publicapicontroller Getjobresult"
                }
              },
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Extraction JSON, raw Markdown, or embeddings JSON."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "JOB_NOT_READY"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "PAYLOAD_TOO_LARGE"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Download job result",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/v1/llm/profiles": {
      "get": {
        "operationId": "PublicLlmProfilesController_listProfiles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List LLM profiles",
        "tags": [
          "LLM"
        ]
      }
    }
  },
...
}